Update README.md
i'm removing the "Unseen Tasks" section because this is a very broad area of study and I can't possibly keep track of the (almost daily) papers in this area.  I think focussing best research on the ultra-high success rate via fine tuning is a better singular goal for this repo.